Finished Text and Updated Clear
noticed the way default things were listed and adjusted clear to fit this to keep continuity.

Text was overhauled to correctly reflect the new perams "small", "medium", and "large, and even added a`custom usage` section that goes over how to call a custom font.

Planning to revamp Font to add more indepth guide of how to setup a Font folder and add your own custom font in!